Started creating gameshell system
GameShell.run(screen) will run the game shell starting with screen. A screen is
just a callable that takes a GameShell instance to use for IO and returns the
next screen to run. This effectively makes the GameShell a finite state machine
with the screens representing states and the return statements in the screen
function representing the transition rules.

For convenience sake GameShell also keeps a stack of screens. This allows you
to implement subscreens by pushing the current screen onto the stack (using
GameShell.push()) before returning the subscreen as normal. Within the
subscreen you can then return None to have the shell automatically pop the
stack and return to the calling screen without having to keep track of where
the subscreen was called from manually.

I've added a stubbed menu system in menu.py to demonstrate the system.